+++ out/extension.js
@@ -5,25 +5,39 @@ exports.deactivate = exports.activate = void 0;
 // Import the module and reference it with the alias vscode in your code below
 const vscode = require("vscode");
 const data_1 = require("./data");
+var outputChannel = vscode.window.createOutputChannel("html-translator");
+async function syncWithConfig() {
+    outputChannel.appendLine("Reading configuration...");
+    let errs = await (0, data_1.parseConfig)();
+    if (errs) {
+        errs.forEach(outputChannel.appendLine);
+        vscode.window.showErrorMessage("One or more thing went wrong", "See errors").then(v => v && outputChannel.show());
+    }
+    else {
+        outputChannel.appendLine("Everything is all good to go");
+    }
+    outputChannel.appendLine("");
+}
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 function activate(context) {
-    (0, data_1.syncWithConfiguration)();
-    vscode.window.visibleTextEditors.forEach(editor => (0, data_1.loadTranslationsFor)(editor.document));
+    // syncWithConfig();
     const configChange = vscode.workspace.onDidChangeConfiguration((e) => {
         if (!e.affectsConfiguration("html-translator"))
             return;
         // if(e.affectsConfiguration("html-translator.languages")) 
         // if(e.affectsConfiguration("html-translator.files")) 
-        (0, data_1.syncWithConfiguration)();
+        syncWithConfig();
     });
-    const openDocumentDisposable = vscode.workspace.onDidOpenTextDocument(data_1.loadTranslationsFor);
-    const closeDocumentDisposable = vscode.workspace.onDidCloseTextDocument(data_1.unloadTranslationsFor);
+    const start = vscode.commands.registerCommand("html-translator.start", syncWithConfig);
+    const stop = vscode.commands.registerCommand("html-translator.stop", data_1.clearAll);
+    // const openDocumentDisposable = vscode.workspace.onDidOpenTextDocument(loadTranslationsFor);
+    // const closeDocumentDisposable = vscode.workspace.onDidCloseTextDocument(unloadTranslationsFor);
     const translUpdate = vscode.workspace.onDidSaveTextDocument(data_1.updateTranslationsFrom);
     const completition = vscode.languages.registerCompletionItemProvider("html", {
         provideCompletionItems: data_1.getSuggestions
-    }, '.', '{');
-    context.subscriptions.push(configChange, openDocumentDisposable, closeDocumentDisposable, translUpdate, completition);
+    }, '.');
+    context.subscriptions.push(configChange, start, stop, translUpdate, completition);
 }
 exports.activate = activate;
 // this method is called when your extension is deactivated